Update and rename main.js to exp1main.js
